+++ packages/build-tools/utils/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.8.0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.8.0-beta.6...v2.8.0) (2019-10-11)
+
+**Note:** Version bump only for package @bolt/build-utils
+
+
+
+
+
+# [2.8.0-beta.4](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.8.0-beta.3...v2.8.0-beta.4) (2019-09-26)
+
+
+### Bug Fixes
+
+* address eslint issues ([e90e7b3](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/e90e7b3))
+
+
+### Features
+
+* automatically convert Bolt + non-Bolt package.json package names into valid Twig namespaces ([bbb6184](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/bbb6184))
+
+
+
+
+
+# [2.8.0-beta.3](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.7.1...v2.8.0-beta.3) (2019-09-24)
+
+
+
+# [2.8.0-beta.2](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.7.0...v2.8.0-beta.2) (2019-09-18)
+
+
+### Bug Fixes
+
+* add .js AND ref support schemas ([300faab](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/300faab))
+
+
+### Features
+
+* **build-tools:** add JS as schema file option ([4f7dbd0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/4f7dbd0))
+
+
+
+
+
+# [2.8.0-beta.2](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.7.0...v2.8.0-beta.2) (2019-09-18)
+
+
+### Bug Fixes
+
+* add .js AND ref support schemas ([300faab](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/300faab))
+
+
+### Features
+
+* **build-tools:** add JS as schema file option ([4f7dbd0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/4f7dbd0))
+
+
+
+
+
+# [2.8.0-beta.1](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.7.0...v2.8.0-beta.1) (2019-09-14)
+
+
+### Bug Fixes
+
+* add .js AND ref support schemas ([300faab](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/300faab))
+
+
+### Features
+
+* **build-tools:** add JS as schema file option ([4f7dbd0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/commit/4f7dbd0))
+
+
+
+
+
 # [2.7.0](https://github.com/bolt-design-system/bolt/tree/master/packages/build-tools/utils/compare/v2.6.0...v2.7.0) (2019-09-13)