+++ packages/core/CHANGELOG.md
@@ -3,6 +3,114 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.1](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.8.0...v2.8.1) (2019-10-16)
+
+
+### Features
+
+* add adaptive menu to Tabs component ([5da94fb](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/5da94fb))
+
+
+
+
+
+# [2.8.0](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.8.0-beta.6...v2.8.0) (2019-10-11)
+
+**Note:** Version bump only for package @bolt/core
+
+
+
+
+
+# [2.8.0-beta.6](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.8.0-beta.5...v2.8.0-beta.6) (2019-10-04)
+
+
+### Bug Fixes
+
+* **micro-journeys:** Context JS error when animations have no parent bolt-interactive-pathways w/ context ([44f85b0](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/44f85b0))
+* add Weakset polyfill for IE 11 ([0867e5f](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/0867e5f))
+
+
+### Reverts
+
+* revert combining withLitHTML base with withLitContext for now ([ca828da](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/ca828da))
+
+
+
+
+
+# [2.8.0-beta.4](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.8.0-beta.3...v2.8.0-beta.4) (2019-09-26)
+
+
+### Bug Fixes
+
+* pull in custom Switch JS to fix custom event issue; add example with/without page with Shadow DOM disabled for debugging; ([8261a6f](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/8261a6f))
+
+
+
+
+
+# [2.8.0-beta.3](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.7.1...v2.8.0-beta.3) (2019-09-24)
+
+
+
+# [2.8.0-beta.2](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.7.0...v2.8.0-beta.2) (2019-09-18)
+
+
+### Bug Fixes
+
+* **animate:** IE11 issue by not using Array.from() ([0dfd86c](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/0dfd86c))
+* **dropdown:** add Node.replaceWith polyfill ([6a870cf](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/6a870cf))
+
+
+### Features
+
+* **core:** add build meta data to window.bolt.meta ([1ac2207](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/1ac2207))
+* **core:** add getData JS function for getting "NAME.bolt.json" data ([e3d96e4](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/e3d96e4))
+* **micro-journeys:** re-style tabs after refactor; replace colors with theme ones ([8a2f14e](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/8a2f14e))
+
+
+
+
+
+# [2.8.0-beta.2](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.7.0...v2.8.0-beta.2) (2019-09-18)
+
+
+### Bug Fixes
+
+* **animate:** IE11 issue by not using Array.from() ([0dfd86c](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/0dfd86c))
+* **dropdown:** add Node.replaceWith polyfill ([6a870cf](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/6a870cf))
+
+
+### Features
+
+* **core:** add build meta data to window.bolt.meta ([1ac2207](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/1ac2207))
+* **core:** add getData JS function for getting "NAME.bolt.json" data ([e3d96e4](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/e3d96e4))
+* **micro-journeys:** re-style tabs after refactor; replace colors with theme ones ([8a2f14e](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/8a2f14e))
+
+
+
+
+
+# [2.8.0-beta.1](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.7.0...v2.8.0-beta.1) (2019-09-14)
+
+
+### Bug Fixes
+
+* **animate:** IE11 issue by not using Array.from() ([0dfd86c](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/0dfd86c))
+* **dropdown:** add Node.replaceWith polyfill ([6a870cf](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/6a870cf))
+
+
+### Features
+
+* **core:** add build meta data to window.bolt.meta ([1ac2207](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/1ac2207))
+* **core:** add getData JS function for getting "NAME.bolt.json" data ([e3d96e4](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/e3d96e4))
+* **micro-journeys:** re-style tabs after refactor; replace colors with theme ones ([8a2f14e](https://github.com/bolt-design-system/bolt/tree/master/packages/core/commit/8a2f14e))
+
+
+
+
+
 # [2.7.0](https://github.com/bolt-design-system/bolt/tree/master/packages/core/compare/v2.6.0...v2.7.0) (2019-09-13)